+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Added `GeminiURI.without_query`.
   ([#32](https://github.com/davep/wasat/pull/32))
 - Added `GeminiURI.parent` ([#32](https://github.com/davep/wasat/pull/32))
+- Added `GeminiURI.root` ([#32](https://github.com/davep/wasat/pull/32))
 
 ## v0.7.0